Add Spark driver for finding nearest neighbors from topic vectors
This driver is intended to consume the output of the topic model, and
use it to generate sparse topic vectors representing the dataset
descriptions. Each topic vector contains the top 10% of the topics for
the corresponding description. Similar items (by cosine distance) are
identified using an approximate nearest neighbors approach based on
sign-random-projection locality-sensitive hashing.